Move deeplink module to core
