+++ app/README.md
@@ -39,6 +39,11 @@ graph TB
     :core:connectivity:api[api]:::multiplatform
     :core:connectivity:implementation[implementation]:::multiplatform
   end
+  subgraph :core:deeplink
+    direction TB
+    :core:deeplink:api[api]:::multiplatform
+    :core:deeplink:implementation[implementation]:::multiplatform
+  end
   subgraph :core:feature-flags
     direction TB
     :core:feature-flags:api[api]:::multiplatform
@@ -281,11 +286,6 @@ graph TB
     :data:watchproviders:api[api]:::multiplatform
     :data:watchproviders:implementation[implementation]:::multiplatform
   end
-  subgraph :deeplink
-    direction TB
-    :deeplink:api[api]:::multiplatform
-    :deeplink:implementation[implementation]:::multiplatform
-  end
   subgraph :domain
     direction TB
     :domain:account-switcher[account-switcher]:::multiplatform
@@ -521,6 +521,8 @@ graph TB
   :app -.-> :core:appconfig:implementation
   :app -.-> :core:base
   :app -.-> :core:connectivity:implementation
+  :app -.-> :core:deeplink:api
+  :app -.-> :core:deeplink:implementation
   :app -.-> :core:feature-flags:api
   :app -.-> :core:feature-flags:implementation
   :app -.-> :core:imageloading:implementation
@@ -607,8 +609,6 @@ graph TB
   :app -.-> :data:watchlist-prefs:implementation
   :app -.-> :data:watchproviders:api
   :app -.-> :data:watchproviders:implementation
-  :app -.-> :deeplink:api
-  :app -.-> :deeplink:implementation
   :app -.-> :domain:account-switcher
   :app -.-> :domain:backup
   :app -.-> :domain:calendar
@@ -712,6 +712,7 @@ graph TB
   :core:base --> :core:view
   :core:connectivity:implementation -.-> :core:base
   :core:connectivity:implementation --> :core:connectivity:api
+  :core:deeplink:implementation --> :core:deeplink:api
   :core:feature-flags:implementation --> :core:appconfig:api
   :core:feature-flags:implementation --> :core:base
   :core:feature-flags:implementation --> :core:feature-flags:api
@@ -1104,7 +1105,6 @@ graph TB
   :data:watchproviders:implementation --> :data:database:sqldelight
   :data:watchproviders:implementation --> :data:request-manager:api
   :data:watchproviders:implementation --> :data:watchproviders:api
-  :deeplink:implementation --> :deeplink:api
   :domain:account-switcher --> :core:base
   :domain:account-switcher --> :core:logger:api
   :domain:account-switcher --> :data:account-manager:api
@@ -1549,12 +1549,12 @@ graph TB
   :features:root:nav --> :domain:theme
   :features:root:presenter --> :core:base
   :features:root:presenter --> :core:connectivity:api
+  :features:root:presenter --> :core:deeplink:api
   :features:root:presenter --> :core:logger:api
   :features:root:presenter --> :core:syncstate:api
   :features:root:presenter -.-> :core:view
   :features:root:presenter --> :data:account-manager:api
   :features:root:presenter --> :data:datastore:api
-  :features:root:presenter --> :deeplink:api
   :features:root:presenter --> :domain:episode
   :features:root:presenter --> :domain:logout
   :features:root:presenter -.-> :domain:theme
