+++ core/deeplink/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/core/deeplink/api/DeepLinkParser.kt
@@ -1,0 +1,5 @@
+package com.thomaskioko.tvmaniac.core.deeplink.api
+
+public interface DeepLinkParser {
+    public fun parse(url: String?): DeepLink?
+}
