+++ core/deeplink/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/core/deeplink/api/DeepLinkUrls.kt
@@ -1,0 +1,5 @@
+package com.thomaskioko.tvmaniac.core.deeplink.api
+
+public interface DeepLinkUrls {
+    public fun urlFor(deepLink: DeepLink): String?
+}
