+++ core/integration/infra/README.md
@@ -35,6 +35,11 @@ graph TB
     :core:connectivity:api[api]:::multiplatform
     :core:connectivity:implementation[implementation]:::multiplatform
   end
+  subgraph :core:deeplink
+    direction TB
+    :core:deeplink:api[api]:::multiplatform
+    :core:deeplink:implementation[implementation]:::multiplatform
+  end
   subgraph :core:feature-flags
     direction TB
     :core:feature-flags:api[api]:::multiplatform
@@ -230,11 +235,6 @@ graph TB
     :data:watchproviders:api[api]:::multiplatform
     :data:watchproviders:implementation[implementation]:::multiplatform
   end
-  subgraph :deeplink
-    direction TB
-    :deeplink:api[api]:::multiplatform
-    :deeplink:implementation[implementation]:::multiplatform
-  end
   subgraph :domain
     direction TB
     :domain:account-switcher[account-switcher]:::multiplatform
@@ -364,6 +364,7 @@ graph TB
   :core:base --> :core:view
   :core:connectivity:implementation -.-> :core:base
   :core:connectivity:implementation --> :core:connectivity:api
+  :core:deeplink:implementation --> :core:deeplink:api
   :core:feature-flags:implementation --> :core:appconfig:api
   :core:feature-flags:implementation --> :core:base
   :core:feature-flags:implementation --> :core:feature-flags:api
@@ -379,6 +380,8 @@ graph TB
   :core:integration:infra --> :core:base
   :core:integration:infra --> :core:connectivity:api
   :core:integration:infra --> :core:connectivity:implementation
+  :core:integration:infra --> :core:deeplink:api
+  :core:integration:infra --> :core:deeplink:implementation
   :core:integration:infra --> :core:feature-flags:api
   :core:integration:infra --> :core:feature-flags:implementation
   :core:integration:infra --> :core:integration:stubs
@@ -421,8 +424,6 @@ graph TB
   :core:integration:infra --> :data:user:implementation
   :core:integration:infra --> :data:watch-status:implementation
   :core:integration:infra --> :data:watchproviders:implementation
-  :core:integration:infra --> :deeplink:api
-  :core:integration:infra --> :deeplink:implementation
   :core:integration:infra --> :domain:backup
   :core:integration:infra --> :domain:continue-watching
   :core:integration:infra --> :domain:episode
@@ -692,7 +693,6 @@ graph TB
   :data:watchproviders:implementation --> :data:database:sqldelight
   :data:watchproviders:implementation --> :data:request-manager:api
   :data:watchproviders:implementation --> :data:watchproviders:api
-  :deeplink:implementation --> :deeplink:api
   :domain:account-switcher --> :core:base
   :domain:account-switcher --> :core:logger:api
   :domain:account-switcher --> :data:account-manager:api
@@ -818,12 +818,12 @@ graph TB
   :features:root:nav --> :domain:theme
   :features:root:presenter --> :core:base
   :features:root:presenter --> :core:connectivity:api
+  :features:root:presenter --> :core:deeplink:api
   :features:root:presenter --> :core:logger:api
   :features:root:presenter --> :core:syncstate:api
   :features:root:presenter -.-> :core:view
   :features:root:presenter --> :data:account-manager:api
   :features:root:presenter --> :data:datastore:api
-  :features:root:presenter --> :deeplink:api
   :features:root:presenter --> :domain:episode
   :features:root:presenter --> :domain:logout
   :features:root:presenter -.-> :domain:theme
