--- deeplink/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/deeplink/api/DeepLinkParser.kt
@@ -1,5 +1,0 @@
-package com.thomaskioko.tvmaniac.deeplink.api
-
-public interface DeepLinkParser {
-    public fun parse(url: String?): DeepLink?
-}
