--- deeplink/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/deeplink/api/DeepLinkUrls.kt
@@ -1,5 +1,0 @@
-package com.thomaskioko.tvmaniac.deeplink.api
-
-public interface DeepLinkUrls {
-    public fun urlFor(deepLink: DeepLink): String?
-}
