--- deeplink/implementation/README.md
@@ -1,42 +1,0 @@
-# `:deeplink:implementation`
-
-## Module dependency graph
-
-<!--region graph-->
-```mermaid
-graph TB
-  subgraph :deeplink
-    direction TB
-    :deeplink:api[api]:::multiplatform
-    :deeplink:implementation[implementation]:::multiplatform
-  end
-
-  :deeplink:implementation --> :deeplink:api
-
-classDef application fill:#CAFFBF,stroke:#000,stroke-width:2px,color:#000;
-classDef multiplatform fill:#FFD6A5,stroke:#000,stroke-width:2px,color:#000;
-classDef android-library fill:#9BF6FF,stroke:#000,stroke-width:2px,color:#000;
-classDef jvm-library fill:#BDB2FF,stroke:#000,stroke-width:2px,color:#000;
-classDef unknown fill:#FFADAD,stroke:#000,stroke-width:2px,color:#000;
-```
-
-<details><summary>Graph legend</summary>
-
-```mermaid
-graph TB
-  application[application]:::application
-  multiplatform[multiplatform]:::multiplatform
-  android-library[android-library]:::android-library
-  jvm-library[jvm-library]:::jvm-library
-
-  api["api dependency"] --> implementation["implementation dependency"]
-
-classDef application fill:#CAFFBF,stroke:#000,stroke-width:2px,color:#000;
-classDef multiplatform fill:#FFD6A5,stroke:#000,stroke-width:2px,color:#000;
-classDef android-library fill:#9BF6FF,stroke:#000,stroke-width:2px,color:#000;
-classDef jvm-library fill:#BDB2FF,stroke:#000,stroke-width:2px,color:#000;
-classDef unknown fill:#FFADAD,stroke:#000,stroke-width:2px,color:#000;
-```
-
-</details>
-<!--endregion-->
