--- deeplink/implementation/build.gradle.kts
@@ -1,23 +1,0 @@
-plugins {
-    alias(libs.plugins.app.kmp)
-}
-
-scaffold {
-    useMetro()
-}
-
-kotlin {
-    sourceSets {
-        commonMain {
-            dependencies {
-                api(projects.deeplink.api)
-            }
-        }
-
-        commonTest {
-            dependencies {
-                implementation(libs.bundles.unittest)
-            }
-        }
-    }
-}
