--- deeplink/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/deeplink/implementation/DeepLinkGrammar.kt
@@ -1,9 +1,0 @@
-package com.thomaskioko.tvmaniac.deeplink.implementation
-
-internal object DeepLinkGrammar {
-    const val SCHEME: String = "tvmaniac"
-    const val HOST_SHOW: String = "show"
-    const val HOST_EPISODE: String = "episode"
-    const val SHOW_SEGMENTS: Int = 1
-    const val EPISODE_SEGMENTS: Int = 3
-}
