--- deeplink/testing/build.gradle.kts
@@ -1,13 +1,0 @@
-plugins {
-    alias(libs.plugins.app.kmp)
-}
-
-kotlin {
-    sourceSets {
-        commonMain {
-            dependencies {
-                api(projects.deeplink.api)
-            }
-        }
-    }
-}
