+++ ios-framework/build.gradle.kts
@@ -68,7 +68,7 @@ scaffold {
             export(projects.features.watchdateSelection.presenter)
             export(projects.features.featureFlags.presenter)
             export(projects.features.root.presenter)
-            export(projects.deeplink.api)
+            export(projects.core.deeplink.api)
             export(projects.features.root.nav)
             export(projects.core.testTags)
 
@@ -125,8 +125,8 @@ kotlin {
                 api(projects.features.root.presenter)
                 implementation(projects.features.seasonDetails.nav)
                 api(projects.features.showDetails.nav)
-                api(projects.deeplink.api)
-                api(projects.deeplink.implementation)
+                api(projects.core.deeplink.api)
+                api(projects.core.deeplink.implementation)
                 api(projects.features.root.nav)
                 api(projects.features.discover.nav)
                 api(projects.features.progress.nav)
