+++ settings.gradle.kts
@@ -119,6 +119,9 @@ include(
     ":core:connectivity:api",
     ":core:connectivity:implementation",
     ":core:connectivity:testing",
+    ":core:deeplink:api",
+    ":core:deeplink:implementation",
+    ":core:deeplink:testing",
     ":core:feature-flags:api",
     ":core:feature-flags:implementation",
     ":core:feature-flags:testing",
@@ -302,9 +305,6 @@ include(
     ":i18n:generator",
     ":i18n:implementation",
     ":i18n:testing",
-    ":deeplink:api",
-    ":deeplink:implementation",
-    ":deeplink:testing",
     ":navigation:api",
     ":navigation:implementation",
     ":navigation:testing",
